requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
dataclasses==0.6
Django==2.2.16
django-cors-headers==3.6.0
django-environ==0.4.5
django-mongoengine==0.4.1
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-resized==0.3.11
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangorestframework-simplejwt==2.1
djongo==1.3.1
idna==3.3
mongoengine==0.19.1
phonenumbers==8.12.39
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.3
pytz==2021.3
requests==2.26.0
sendgrid==6.9.1
six==1.15.0
sqlparse==0.2.4
starkbank-ecdsa==2.0.3
twilio==7.3.1
urllib3==1.26.7
gunicorn==20.1.0
psycopg2-binary==2.9.5